requirements.txt
coreapi
django
djangorestframework
django-filter
markdown
mysql-python==1.2.5
pymongo
paho-mqtt